UserDataInsertion(Driver, MyConnection) Feature #1
- Added code to the driver class to give the user the option to insert
data.
- Added the proper insertion SQL code in MyConnection, which is called
by the Driver class, to allow the user to input data. The SQL code runs
as a transaction that can be rolledback, due to there being multiple
insertions into multiple tables.
- The Database schema has been slightly revised to better store data for
the source of a word. The source is now part of the primary key with
word id, and the code in MyConnection reflects this change.