LocalWords(Driver,LocalWord,MyConnection) Feature #6 #9 #11
- Added delete command to Driver class, deletes a word based on wid via
the MyConnection class's removeWord method.
- Added removeWord method to MyConnection, removes a word based on wid
and language.
- Added listhomonyms command to Driver class, lists all the homonyms of
a word, based on the word and its language.
- Added listHomonyms method to MyConnection, facilitates the listing of
database homonyms.
- Implemented pull function in LocalWord class to allow words with
partial data to pull a complete copy of their data from the database.
- Implemented three setvalue functions in LocalWord to set the parts of
its data that relate to the three tables stored in the database: Word,
Symbols and Source.
- tname is now sname due to a naming change in the database schema.
sname represents the source name. tname represented the same data, but
came from a deprecated table called title that linked to the source
table via a source id (which also no longer exists).
- getRowCount is now deprecated as we no longer need to count the rows
of word to determine a new word id. The database now uses auto_increment
to determine new word id.
- Added setvalue functions to MyConnections that correspond to the three
setvalue functions in localWord, and are infact called by the localword
versions of the functions. The setvalue functions in MyConnections
contain all the sql logic used to obtain the database side values, and
then set the values in the LocalWord.