LanguageTests(+Test, +QuestionFactory, +Logger, ...) Feature #10 #7 #3
- Added new classes Test, QuestionFactory, Logger,
SymbolsMainQuestionFactory.
- The Test class is a way of storing and then serving questions to
users, it utilizes a QuestionFactory to create questions.
- The QuestionFactory class is Abstract and must be extended. It is
meant to decouple the way questions are created from the tests they are
used in.
- The SymbolsMainQuestionFactory is question factory that creates
questions about main symbols, with the answers being the ancillary
symbols.
- The Logger class was created to decouple the logging message from the
method of logging. This may be useful if I implement a GUI in the
future.
- The Driver class has been updated to use logging statements instead of
print statements.
- The Driver class had had a create command added to its command list.
The create command currently generates a new test that tests the main
symbols against the ancillary symbols answers.
- The MyConnection class has been updated to use the new Logger class to
log its data instead of using print statements.
- The MyConnection class is now a proper singleton, while in the past it
was not technically a singleton.
- Added getWords function to the MyConnection class that returns a list
of LocalWords based on an SQL command. The SQL command must be in a
seven string return format, however some of the strings may be set as
NULL values.
- The LocalWord class now has its main and ancillary strings available
via getters, due to the new main-ancillary questions requiring access to
that data.
- The Question class now has a method correctAnswer which tests for
question correctness, this facilitates Test's abilities to tell the
using if they have correctly answered a question.