TestCreation (SymbolsMainQuestionFactory, MeaningQuestionFactory..) #13
- Question Factories now have their queries modified via a string passed
to their constructors. This is after complete cuts to attempted
decorator classes. I do not believe that the pattern works in this case,
or I am perhaps not experienced enough to implement it.

- Added MeaningQuestionFactory class, which generates questions about
the meaning of a word's romanization.

- Slight formatting on the Test class. Added a check to see if the
questions list contained zero questions.

- LocalWord now provides access to the values of meaning and
romanization via getter functions.

- Driver now has some SQL code included, when creating question
factories, which I had wished to avoid. I may create a class to
encapsulate SQL modification strings, passing values such as wtype and
source in the constructor, which I would then use the .toString() of
that class when creating a question factory.